feat: add pytest
