+++ README.md
@@ -32,15 +32,15 @@ pip install nerif
 from nerif.core import nerif
 from nerif.model import SimpleChatModel
 
-agent = SimpleChatModel()
+model = SimpleChatModel()
 
 # Use nerif judge "natural language statement"
 if nerif("the sky is blue"):
     print("True")
 else:
-    # Call a simple agent
+    # Call a simple model
     print("No", end=", ")
-    print(agent.chat("what is the color of the sky?"))
+    print(model.chat("what is the color of the sky?"))
 ```
 
 ## Documentation
