+++ test/nerif_format_test.py
@@ -5,7 +5,7 @@ import warnings
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from nerif.core import (
+from nerif.utils import (
     FormatVerifierFloat,
     FormatVerifierHumanReadableList,
     FormatVerifierInt,
