+++ test/nerif_test.py
@@ -4,33 +4,23 @@ import unittest
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from nerif.core import core, nerif_match, nerif_match_string
+from nerif.core import nerif, nerif_match_string
 
 
 class MyTestCase(unittest.TestCase):
     def test_judge(self):
-        judge = core("the sky is blue")
+        judge = nerif("the sky is blue")
         self.assertEqual(True, judge)  # add assertion here
-        judge = core("Do you know who I am?")
+        judge = nerif("Do you know who I am?")
         self.assertEqual(False, judge)
 
     def test_match(self):
-        test_selection = {
-            "1": "the sky is blue",
-            "2": "the sky is green",
-            "3": "the sky is red",
-        }
-        match_result = nerif_match_string("the truth", test_selection)
+        selections = ["iPhone 5", "iPhone 6", "iPhone 12"]
 
-        self.assertEqual("1", match_result)
-        test_selection = {
-            "reboot": "restart the server, it may takes a few minutes",
-            "adduser": "add a new user on my server",
-            "reservation": "make a reservation for the server",
-        }
-        match_result = nerif_match("I wanna use the server for AI training tonight", test_selection)
+        best_choice = nerif_match_string(selections=selections, text="Which iPhone is the most powerful one?")
 
-        self.assertEqual("reservation", match_result)
+        self.assertEqual(2, best_choice)
+
 
 
 if __name__ == "__main__":
